+++ src/dbhose_airflow/core/__init__.py
@@ -1,14 +1,40 @@
 """Rust module for get DDLs from servers."""
 
+from . import errors as Error
+from .ddl import generate_ddl
 from .ddl_core import (
     clickhouse_ddl,
     postgres_ddl,
     postgres_sequence_ddl,
 )
+from .structs import (
+    ConnectionConfig,
+    ColumnMeta,
+    DQCheck,
+    DQConfig,
+    DQTest,
+    ETLInfo,
+    MoveMethod,
+    MoveType,
+    StagingConfig,
+    TableMetadata,
+)
 
 
 __all__ = (
+    "ConnectionConfig",
+    "ColumnMeta",
+    "DQCheck",
+    "DQConfig",
+    "DQTest",
+    "Error",
+    "ETLInfo",
+    "MoveMethod",
+    "MoveType",
+    "StagingConfig",
+    "TableMetadata",
     "clickhouse_ddl",
+    "generate_ddl",
     "postgres_ddl",
     "postgres_sequence_ddl",
 )